Create json interface || refactor class to module
